requirements.txt
functions-framework==3.*
requests
pandas
validators
google-cloud-storage
gcsfs